rename mock epoll to match other naming styles
